+++ products/jbrowse-web/src/tests/__image_snapshots__/reload-test-tsx-reloads-alignments-track-bai-404-1-snap.png
