+++ products/jbrowse-web/src/tests/__image_snapshots__/reload-test-tsx-reloads-alignments-track-crai-404-1-snap.png
