+++ products/jbrowse-web/src/tests/__image_snapshots__/reload-test-tsx-reloads-bigwig-bw-404-1-snap.png
