+++ products/jbrowse-web/src/tests/__image_snapshots__/stats-estimation-test-tsx-test-stats-estimation-pileup-zoom-in-to-see-1-snap.png
